+++ TrueTrace/DemoScene.asset
@@ -31,7 +31,7 @@ MonoBehaviour:
   PPToneMap: 0
   PPTAA: 0
   RenderScale: 1
-  UseASVGF: 0
+  UseASVGF: 1
   UseTAAU: 0
   ReSTIRGIUpdateRate: 7
   UseReSTIRGITemporal: 1
@@ -47,7 +47,7 @@ MonoBehaviour:
   DoFirefly: 0
   ImprovedPrimaryHit: 1
   RISCount: 12
-  ToneMapper: 1
+  ToneMapper: 0
   SkyDesaturate: 0
   ClayColor: {x: 0.20754719, y: 0.20754719, z: 0.20754719}
   GroundColor: {x: 0.11320752, y: 0.11320752, z: 0.11320752}
@@ -56,7 +56,7 @@ MonoBehaviour:
   FireflyStrength: 1
   FireflyOffset: 0
   OIDNFrameCount: 50
-  UseOIDN: 1
+  UseOIDN: 0
   DoSharpen: 0
   Sharpness: 1
   MainDesiredRes: 16384
@@ -75,3 +75,10 @@ MonoBehaviour:
   MatChangeResetsAccum: 0
   PPFXAA: 0
   OIDNBlendRatio: 1
+  ConvBloom: 1
+  ConvStrength: 2.51
+  ConvBloomThreshold: 0
+  ConvBloomSize: {x: 1, y: 1}
+  ConvBloomDistExp: 0
+  ConvBloomDistExpClampMin: 1
+  ConvBloomDistExpScale: 1
