+++ TrueTrace/Resources/MainCompute/RayGenKernels.compute
@@ -79,7 +79,7 @@ void GeneratePanorama(uint3 id : SV_DispatchThreadID) {
     const uint pixel_index = id.y * screen_width + id.x;
     float2 jitter = ((random(0, pixel_index) - 0.5));
     if(OIDNGuideWrite) jitter = 0;
-    float2 uv = float2((id.xy) / float2(screen_width, screen_height));
+    float2 uv = float2((id.xy + jitter) / float2(screen_width, screen_height));
     SmallerRay ray;
     if(DoPanorama) {
         uv.y = 1.0f - uv.y;
