+++ TrueTrace/Resources/PostProcess/FFTBloom.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 937921310f57e6849b283917fe90eaf1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
