+++ TrueTrace/Resources/PostProcess/FFTBloom/ConvolutionBloom.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3e2fcdbece36ea54b8553f382ebf1c8a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
