+++ TrueTrace/Resources/PostProcess/FFTBloom/FFTCS.compute.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3601c6eb74769c140a7c39e53c3a2bfe
+ComputeShaderImporter:
+  externalObjects: {}
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
