+++ TrueTrace/Resources/PostProcess/FFTBloom/FFTCommon.cginc.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 627bb42178520c44680563847a422920
+ShaderIncludeImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
