+++ TrueTrace/Resources/PostProcess/FFTBloom/FinalBlit.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9fd4bb07cd9a7d54089e3e45e7e659dd
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
