+++ TrueTrace/Resources/PostProcess/FFTBloom/KernelGenerate.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9751606b4a041c64bba0853137ad7b20
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
