+++ TrueTrace/Resources/PostProcess/FFTBloom/SourceGenerate.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 133bf56fa57f4e641a5f71049161d588
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
