+++ TrueTrace/Resources/PostProcess/TTPostProcessing.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73588e4c2b8222d4791f8b70e9bb7579
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
