+++ TrueTrace/Resources/Utility/PanoramaDoer.cs
@@ -95,7 +95,7 @@ void AddResolution(int width, int height, string label)
         public void OnDisable() {
             if(RayMaster.DoChainedImages) RemoveResolution(GetCount() - 1);
         }
-        private void FinalizePanorama() {
+        private void FinalizePanorama(Camera cam) {
             Color[] FinalAtlasData = new Color[FinalAtlasSize.x * FinalAtlasSize.y];
 
             for(int iter = 0; iter < HorizontalSegments; iter++) {
@@ -116,7 +116,7 @@ void AddResolution(int width, int height, string label)
             Texture2D FinalAtlas = new Texture2D(FinalAtlasSize.x, FinalAtlasSize.y);
             FinalAtlas.SetPixels(FinalAtlasData, 0);
             FinalAtlas.Apply();
-            System.IO.File.WriteAllBytes(PlayerPrefs.GetString("ScreenShotPath") + "/" + System.DateTime.Now.ToString("yyyy-MM-dd HH-mm-ss") + ".png", FinalAtlas.EncodeToPNG()); 
+            System.IO.File.WriteAllBytes(PlayerPrefs.GetString("PanoramaPath") + "/" + cam.gameObject.name + ".png", FinalAtlas.EncodeToPNG()); 
         }
         public void Init() {
             RayMaster.SampleCount = 0;
@@ -165,7 +165,7 @@ void AddResolution(int width, int height, string label)
                     if(CurrentSegment == HorizontalSegments) {
                         CurrentSegment = 0;
                         waitedTime = 0;
-                        FinalizePanorama();
+                        FinalizePanorama(Cameras[CurrentCamera]);
                         CurrentCamera++;
                         if(CurrentCamera < Cameras.Length) {
                             Cameras[CurrentCamera - 1].gameObject.SetActive(false);
